Edited types of some ints/longs. Created class that instantiates all the objects with values.

+++ app/src/main/java/com/kevintyang/codeclicker/codeclicker/MoneyCounters.java
@@ -36,7 +36,7 @@ public class MoneyCounters {
         clickValue += newClickValue;
     }
 
-    public static void subtractCostOfUpgrades(int cost){
+    public static void subtractCostOfUpgrades(long cost){
 
         currentMoneyCount = currentMoneyCount - cost;
     }
